Add golint as a presubmit check.
Fix all golint errors. Note that this presubmit check assumes that golint
returns non-zero when a lint warning is found. There is a pull request to add
this functionality as the time of this commit but it is not merged yet in the
main repository. If this PR wouldn't be merged, the fallback will be to check
for output.